backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "powerline-backend"
version = "2.0.0-dev"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "pydantic-settings>=2.6",
    "email-validator>=2.2",
    "bcrypt>=4.0",
    "PyJWT>=2.10",
    "twilio>=9.4",
    "structlog>=24.4",
    "phonenumbers>=8.13",
    "python-multipart>=0.0.18",
    "cloudinary>=1.36",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "fakeredis[aioredis]>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
